requirements.txt
asgiref==3.8.1
dj-database-url==2.2.0
Django==5.1
django-filter==24.3
djangorestframework==3.15.2
numpy==2.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sqlparse==0.5.1
typing_extensions==4.12.2
